Fix warning

+++ APSICommon/apsi/tools/interpolate.cpp
@@ -98,7 +98,7 @@ namespace apsi
         result.set(0, size-1, divided_differences[0]);
         for (size_t i = 1; i < size; i++)
         {
-            for (int j = i - 1; j >= 0; j--)
+            for (int j = static_cast<int>(i - 1); j >= 0; j--)
             {
                 result.set(j + 1, j, result);
             }
